agents_files/Clean_Hiring_System/passport_agent/requirements.txt
langchain-openai>=0.2.0
langchain-ollama>=0.2.0
pydantic>=2.9.0
python-dotenv==1.0.1
requests==2.32.5
 cryptography>=41.0.0
